__init__.py
from PytestWithAllure import api

from PytestWithAllure.api import (JWT_SECRET, UserIn_Pydantic, User_Pydantic,
                                  app, authenticate_user, create_user,
                                  fast_api, generate_token, get_current_user,
                                  get_user, oauth2_scheme,)

__all__ = ['JWT_SECRET', 'UserIn_Pydantic', 'User_Pydantic', 'api', 'app',
           'authenticate_user', 'create_user', 'fast_api', 'generate_token',
           'get_current_user', 'get_user', 'oauth2_scheme']

