requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
Faker==16.6.1
fastapi==0.89.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
iniconfig==2.0.0
iso8601==1.1.0
jsonpath==0.82
mkinit==1.0.0
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
priority==2.0.0
pydantic==1.10.4
PyJWT==2.6.0
pypika-tortoise==0.1.6
pytest==7.2.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
toml==0.10.2
tomli==2.0.1
tortoise-orm==0.19.3
typing_extensions==4.4.0
ubelt==1.2.3
urllib3==1.26.14
uvicorn==0.20.0
wsproto==1.2.0
